+++ src/fhir/sdc-import-common.js
@@ -1,5 +1,3 @@
-var LForms = require('../lforms-index');
-
 /**
  *  Defines SDC import functions that are the same across the different FHIR
  *  versions.  The function takes SDC namespace object defined in the sdc export
@@ -1102,13 +1100,21 @@ function addCommonSDCImportFns(ns) {
             pendingPromises.push(fetch(expURL).then(function(response) {
               return response.json();
             }).then(function(parsedJSON) {
-              answers = self.answersFromVS(parsedJSON);
-              if (answers) {
-                LForms._valueSetAnswerCache[expURL] = answers;
-                item.answers = answers;
-                lfData._updateAutocompOptions(item, true);
+              if (parsedJSON.resourceType==="OperationOutcome" ) {
+                var errorOrFatal = parsedJSON.issue.find(item => item.severity==="error" || item.severity==="fatal")
+                if (errorOrFatal) {
+                  throw new Error(errorOrFatal.diagnostics)
+                }
               }
-            }, function fail() {
+              else {
+                answers = self.answersFromVS(parsedJSON);
+                if (answers) {
+                  LForms._valueSetAnswerCache[expURL] = answers;
+                  item.answers = answers;
+                  lfData._updateAutocompOptions(item, true);
+                }  
+              }
+            }).catch(function(error) {              
               throw new Error("Unable to load ValueSet from "+expURL);
             }));
           }
@@ -1124,7 +1130,7 @@ function addCommonSDCImportFns(ns) {
                 item.answers = answers;
                 lfData._updateAutocompOptions(item, true);
               }
-            }, function fail() {
+            }).catch(function(error) {
               throw new Error("Unable to load ValueSet "+item.answerValueSet+ " from FHIR server");
             }));
           }